Move GUI part to a separate file
