+++ src/main.rs
@@ -1,62 +1,10 @@
-use crossbeam_channel::{select, unbounded, Receiver};
-use crossterm::{
-    event::{DisableMouseCapture, EnableMouseCapture, Event, KeyCode},
-    execute,
-    terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
-};
-use std::{io, thread};
-use tui::{
-    backend::CrosstermBackend,
-    widgets::{Block, Borders},
-    Terminal,
-};
+mod main_app;
 
-fn setup_ui_events() -> Receiver<Event> {
-    let (sender, receiver) = unbounded();
-    thread::spawn(move || loop {
-        sender.send(crossterm::event::read().unwrap()).unwrap();
-    });
-
-    receiver
-}
+use main_app::App;
+use std::io;
 
 fn main() -> Result<(), io::Error> {
-    // setup terminal
-    enable_raw_mode()?;
-    let mut stdout = io::stdout();
-    execute!(stdout, EnterAlternateScreen, EnableMouseCapture)?;
-    let backend = CrosstermBackend::new(stdout);
-    let mut terminal = Terminal::new(backend)?;
+    let mut app = App::new();
 
-    let ui_events_receiver = setup_ui_events();
-    terminal.draw(|f| {
-        let size = f.size();
-        let block = Block::default().title("Block").borders(Borders::ALL);
-        f.render_widget(block, size);
-    })?;
-
-    loop {
-        select! { recv(ui_events_receiver) -> message => {
-
-                if let Event::Key(key_event) = message.unwrap() {
-                    if key_event.modifiers.is_empty() {
-                            if let KeyCode::Char('q') = key_event.code {
-                                break
-                            }
-                    }
-                }
-        }
-        }
-    }
-
-    // restore terminal
-    disable_raw_mode()?;
-    execute!(
-        terminal.backend_mut(),
-        LeaveAlternateScreen,
-        DisableMouseCapture
-    )?;
-    terminal.show_cursor()?;
-
-    Ok(())
+    app.run()
 }
